% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ncim_int.R
\name{add_indexes}
\alias{add_indexes}
\title{Add Indexes}
\usage{
add_indexes(conn, schema, verbose = TRUE, render_sql = TRUE)
}
\description{
Add Indexes
}
\seealso{
\code{\link[progress]{progress_bar}}
\code{\link[SqlRender]{render}}
\code{\link[pg13]{send}}
\code{\link[secretary]{typewrite_warning}},\code{\link[secretary]{c("typewrite", "typewrite")}},\code{\link[secretary]{italicize}}
\code{\link[purrr]{map}}
}
