% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_mrhier.R
\name{copy_mrhier}
\alias{copy_mrhier}
\title{MRHIER}
\usage{
copy_mrhier(path_to_rrfs, conn, schema, verbose = TRUE, render_sql = TRUE)
}
\description{
Due to perpetual errors, MRHIER is processed in a
roundabout manner. This is because some rows contain
an extra unknown column. In the script, anything
with an index of 11 or greater is removed.
}
\seealso{
\code{\link[pg13]{lsSchema}},\code{\link[pg13]{send}}
\code{\link[SqlRender]{render}}
}
