% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_owl.R
\name{download_owl}
\alias{download_owl}
\title{Download and Extract NCIt OWL Files}
\usage{
download_owl(owl_folder = "/Users/mpatel/terminology/NCIT")
}
\description{
If a diff is detected at the FTP site involving the
Thesaurus OWL files, the new versions will be downloaded,
extracted, and reported to the console.
}
