% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_code_cui_map.R
\name{load_code_cui_map}
\alias{load_code_cui_map}
\title{Load the Code to CUI Map}
\usage{
load_code_cui_map(
  conn,
  schema = "nci",
  log_schema = "public",
  log_table = "setup_nci_log"
)
}
\arguments{
\item{conn}{Postgres connection object.}

\item{schema}{Schema to write to, Default: 'nci'}

\item{log_schema}{Schema for the log table, Default: 'public'}

\item{log_table}{Name of log table, Default: 'setup_nci_log'}
}
\description{
The Code to CUI Map is provided as a \code{dat} file is written
to the \code{CODE_CUI_MAP} table in the given schema.
}
\details{
The \code{dat} file contains the map of NCI Thesaurus (NCIt) codes
to NCI Metathesaurus (NCIMeta) CUIs (concept unique
identifier). The CUIs are either derived from the UMLS
Metathesaurus (CUIs that begin with "C-digit") or
specific to the NCI Metathesaurus (CUIs that begin with
"CL").  This mapping file is created for each release of
the NCIMeta, which is published on a quarterly schedule
hence is not current with every release of NCIt which is
published monthly. Newly created Thesaurus concepts will
not appear in this file until the next release of the
Metathesaurus, and newly retired Thesaurus concepts will
continue to appear until the next release of the
Metathesaurus.
}
\seealso{
\code{\link[xml2]{read_xml}}
\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_table}}
\code{\link[dplyr]{select}},\code{\link[dplyr]{mutate}}
\code{\link[rubix]{format_colnames}},\code{\link[rubix]{filter_at_grepl}}
\code{\link[lubridate]{as_date}}
\code{\link[stringr]{str_replace}}
\code{\link[readr]{read_delim}}
\code{\link[pg13]{schema_exists}},\code{\link[pg13]{create_schema}},\code{\link[pg13]{write_table}},\code{\link[pg13]{table_exists}},\code{\link[pg13]{send}},\code{\link[pg13]{append_table}}
\code{\link[SqlRender]{render}}
\code{\link[tibble]{tibble}}
}
