% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ncim.R
\name{load_ncim}
\alias{load_ncim}
\title{Load NCI Metathesaurus}
\usage{
load_ncim(
  conn,
  schema = "nci",
  path_to_rrfs,
  steps = c("reset_schema", "ddl_tables", "copy_rrfs", "add_indexes", "log"),
  log_schema = "public",
  log_table = "setup_nci_log",
  ncim_version,
  mrconso_only = FALSE,
  omop_only = FALSE,
  english_only = TRUE,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\arguments{
\item{conn}{Postgres connection object.}

\item{schema}{Schema to write to, Default: 'nci'}

\item{log_schema}{Schema for the log table, Default: 'public'}

\item{log_table}{Name of log table, Default: 'setup_nci_log'}

\item{ncim_version}{NCI Metathesaurus version}
}
\description{
Load the NCI Metathesaurus RRF files that are downloaded
from \href{https://evs.nci.nih.gov/evs-download/metathesaurus-downloads}{NIH EVS}.
}
\seealso{
\code{\link[metathesaurus]{setup_pg_mth}}
\code{\link[pg13]{table_exists}},\code{\link[pg13]{send}},\code{\link[pg13]{append_table}}
\code{\link[SqlRender]{render}}
\code{\link[tibble]{tibble}}
}
