% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ncit.R
\name{load_ncit}
\alias{load_ncit}
\title{Load NCI Thesaurus}
\usage{
load_ncit(
  conn,
  schema = "nci",
  log_schema = "public",
  log_table = "setup_nci_log"
)
}
\arguments{
\item{conn}{Postgres connection object.}

\item{schema}{Schema to write to, Default: 'nci'}

\item{log_schema}{Schema for the log table, Default: 'public'}

\item{log_table}{Name of log table, Default: 'setup_nci_log'}
}
\description{
Download the most recent NCIt FLAT file and populate a
\code{THESAURUS} table in a schema.
}
\seealso{
\code{\link[xml2]{read_xml}}
\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_table}}
\code{\link[dplyr]{select}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{desc}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{ranking}}
\code{\link[rubix]{format_colnames}},\code{\link[rubix]{filter_at_grepl}}
\code{\link[lubridate]{as_date}}
\code{\link[stringr]{str_replace}}
\code{\link[utils]{unzip}}
\code{\link[readr]{read_delim}}
\code{\link[pg13]{write_table}},\code{\link[pg13]{table_exists}},\code{\link[pg13]{send}},\code{\link[pg13]{append_table}}
\code{\link[SqlRender]{render}}
\code{\link[tibble]{tibble}}
}
