% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_arguments.R
\name{package_arguments}
\alias{package_arguments}
\title{Package Arguments}
\arguments{
\item{conn}{Postgres connection object.}

\item{schema}{Schema to write to, Default: 'nci'}

\item{log_schema}{Schema for the log table, Default: 'public'}

\item{log_table}{Name of log table, Default: 'setup_nci_log'}
}
\description{
Package Arguments
}
\keyword{internal}
