% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_neo4j_to_omop.R
\name{process_neo4j_to_omop}
\alias{process_neo4j_to_omop}
\title{Process Neo4j files into OMOP Tables}
\usage{
process_neo4j_to_omop(output_folder = file.path(here::here(), "dev"))
}
\description{
Process Neo4j files into OMOP Tables
}
