% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_owl_to_neo4j.R
\name{process_owl_to_neo4j}
\alias{process_owl_to_neo4j}
\title{Process NCI Thesaurus
OWL Files into Nodes and Edges}
\usage{
process_owl_to_neo4j(
  owl_folder,
  output_folder = file.path(here::here(), "dev")
)
}
\arguments{
\item{owl_folder}{Path to the 'Thesaurus.owl' and 'ThesaurusInferred.owl' files downloaded
from the \href{https://evs.nci.nih.gov/ftp1/NCI_Thesaurus/}{NCI Thesaurus FTP}. An error will
be returned if both files cannot be found in the path and if both files are found, if the
NCI Thesaurus versions for both are not the same.}

\item{output_folder}{Destination folder for the outputs generated by this function. Defaults to
\code{dev}.}
}
\value{
New node.csv, edge.csv, and nci_version.txt files in the output_folder/neo4j directory.
}
\description{
Process NCI Thesaurus
OWL Files into Nodes and Edges
}
\details{
Processing occurs via Python and involves JSON intermediaries that are
written to the \code{tmp_neo4j} folder and unlinked on exit.
}
