% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ncim_int.R
\name{reset_schema}
\alias{reset_schema}
\title{Drop All Metathesaurus Tables}
\usage{
reset_schema(conn, schema, verbose = TRUE, render_sql = TRUE)
}
\description{
Drop all the NCI Metathesuarus Tables while preserving the
\verb{Code To CUI Map} and \code{Thesaurus} Tables.
`
}
\seealso{
\code{\link[pg13]{lsSchema}},\code{\link[pg13]{send}}
\code{\link[SqlRender]{render}}
}
