% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-run_setup.R
\name{run_setup}
\alias{run_setup}
\title{Run NCI Setup}
\usage{
run_setup(
  conn,
  conn_fun = "pg13::local_connect()",
  schema = "nci",
  path_to_rrfs,
  ncim_version,
  log_schema = "public",
  log_table = "setup_nci_log"
)
}
\arguments{
\item{conn}{Postgres connection object.}

\item{schema}{Schema to write to, Default: 'nci'}

\item{ncim_version}{NCI Metathesaurus version}

\item{log_schema}{Schema for the log table, Default: 'public'}

\item{log_table}{Name of log table, Default: 'setup_nci_log'}
}
\description{
Load the NCI Metathesaurus RRF files, followed by the
NCI Thesaurus and the Code to CUI Map.
}
