% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ncim.R
\name{setup_ncim}
\alias{setup_ncim}
\title{Instantiate NCI Metathesaurus}
\usage{
setup_ncim(
  conn,
  conn_fun = "pg13::local_connect()",
  schema = "ncim",
  path_to_rrfs,
  steps = c("reset_schema", "ddl_tables", "copy_rrfs", "copy_mrhier", "log",
    "add_indexes"),
  mrconso_only = FALSE,
  omop_only = FALSE,
  english_only = TRUE,
  log_schema = "public",
  log_table_name = "setup_ncim_log",
  log_version,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\description{
Instantiate NCI Metathesaurus
}
