% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ncit.R
\name{setup_ncit}
\alias{setup_ncit}
\title{Run NCI Thesaurus Setup}
\usage{
setup_ncit(
  conn,
  conn_fun = "pg13::local_connect()",
  schema = "ncit",
  log_schema = "public",
  log_table = "setup_ncit_log"
)
}
\arguments{
\item{conn}{Postgres connection object.}

\item{schema}{Schema to write to, Default: 'nci'}

\item{log_schema}{Schema for the log table, Default: 'public'}

\item{log_table}{Name of log table, Default: 'setup_nci_log'}
}
\description{
Load the NCI Thesaurus and the Code to CUI Map.
}
