% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_omop.R
\name{setup_omop}
\alias{setup_omop}
\title{Setup OMOP}
\usage{
setup_omop(
  conn,
  conn_fun = "pg13::local_connect()",
  target_schema = "omop_athena_nci_only",
  steps = c("drop_tables", "copy", "log", "indices", "constraints"),
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  checks = "",
  log_schema = "public",
  log_table = "setup_nci_omop_log"
)
}
\description{
Instantiate a Postgres OMOP Vocabulary
database from the csvs created by the \code{process_owl_to_omop}
function.
}
