% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_owl.R
\name{download_owl}
\alias{download_owl}
\title{Download and Extract NCIt OWL Files}
\usage{
download_owl(owl_folder = "/Users/mpatel/terminology/NCIT")
}
\description{
If a diff is detected at the FTP site involving the
Thesaurus OWL files or the \code{owl_folder} argument and its
file list, the newest version of both NCI Thesaurus OWL files
will be downloaded, extracted, and reported to the console.
}
