% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owl.R
\name{process_owl_to_neo4j}
\alias{process_owl_to_neo4j}
\title{Process NCI Thesaurus
OWL Files into Nodes and Edges}
\usage{
process_owl_to_neo4j(nci_version, output_folder)
}
\arguments{
\item{owl_folder}{Path to the 'Thesaurus.owl' and 'InferredThesaurus.owl' files.}

\item{base_folder}{Base folder to which the json and final
node.csv and edge.csv files are written by \code{nci_version}.}
}
\description{
Process NCI Thesaurus
OWL Files into Nodes and Edges
}
