% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_owl_to_omop.R
\name{process_owl_to_omop}
\alias{process_owl_to_omop}
\title{Process NCI Thesaurus OWL into OMOP Tables}
\usage{
process_owl_to_omop(
  owl_folder,
  output_folder = file.path(here::here(), "dev"),
  inst_folder = file.path(here::here(), "inst", "data")
)
}
\arguments{
\item{owl_folder}{Path to the 'Thesaurus.owl' and 'ThesaurusInferred.owl' files downloaded
from the \href{https://evs.nci.nih.gov/ftp1/NCI_Thesaurus/}{NCI Thesaurus FTP}. An error will
be returned if both files cannot be found in the path and if both files are found, if the
NCI Thesaurus versions for both are not the same.}

\item{output_folder}{Destination folder for the outputs generated by this function. Defaults to
\code{dev}.}

\item{inst_folder}{Installation folder that will include the zipped versions of the Neo4j and OMOP
csvs in deployed packages}
}
\value{
CSV files
}
\description{
Process NCI Thesaurus OWL into OMOP Tables
}
\details{
The OWL files are first processed into the Neo4j nodes and edges csvs
using the \code{process_owl_to_neo4j} function in this package. These csvs are then
processed further into OMOP format Vocabulary table csvs: CONCEPT, CONCEPT_SYNONYM,
CONCEPT_RELATIONSHIP, CONCEPT_ANCESTOR. Metadata tables VOCABULARY, RELATIONSHIP,
and CONCEPT_CLASS are also created.

All concept ids in the output have a pattern of 7-billion.

CONCEPT:
The
\code{concept_class_id} are dependent on the placement of the hierarchy. If it is
the topmost ancestor, it is 'Root' Class. If it is a bottom-most descendant, it is
a 'Leaf' Non-Standard Concept. Otherwise the concept is a 'SubClass' Non-Standard Concept.
If there is any value found within the \code{Concept_Status} field in the nodes csv,
the concept is deprecated in the final concept table.

CONCEPT_RELATIONSHIP:
All relationships (asserted, inherited, and annotation) are used to generate these csvs.
The exact \code{rel_type} from the source edge csv is lost in this processing.

CONCEPT_SYNONYM: All language concepts id are 4180186, only
concept synonyms in the \code{FULL_SYN} source field that did not have a
lowercased match to the \code{Preferred_Name} was used.

CONCEPT_CROSSWALK: New table that contains select crosswalks explicitly stated in
the nodes.csv file.
}
