% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_owl_to_omop.R
\name{process_owl_to_omop}
\alias{process_owl_to_omop}
\title{Process NCI Thesaurus OWL into OMOP Tables}
\usage{
process_owl_to_omop(nci_version)
}
\value{
CSV files
}
\description{
Process NCI Thesaurus OWL into OMOP Tables
}
\details{
The OWL files are first processed into the Neo4j nodes and edges csvs
using the \code{process_owl_to_neo4j} function in this package. These csvs are then
processed further into OMOP format Vocabulary table csvs: CONCEPT, CONCEPT_SYNONYM,
CONCEPT_RELATIONSHIP, CONCEPT_ANCESTOR. Metadata tables VOCABULARY, RELATIONSHIP,
and CONCEPT_CLASS are also created.

CONCEPT:
Concept ids are generated by row number + 20000. Concept ids are not created for
any of the metadata concepts in the VOCABULARY, RELATIONSHIP and CONCEPT_CLASS tables.
The \code{domain_id} is determined by the source \code{rel_type} value in the Edges file. The
\code{concept_class_id} are dependant on the placement of the hierarchy. If it is
the topmost ancestor, it is 'Root' Class. If it is a bottomost descendant, it is
a 'Leaf' Non-Standard Concept. Otherwise the concept is a 'SubClass' Non-Standard Concept.
If there is any value found within the \code{Concept_Status} field in the nodes csv,
the concept is deprecated in the final concept table.

CONCEPT_RELATIONSHIP:
All relationships (asserted, inherited, and annotation) are used to generate these csvs.
The exact \code{rel_type} from the source edge csv is lost in this processing.

CONCEPT_SYNONYM: All language concepts id are 4180186, only
concept synonyms in the \code{FULL_SYN} source field that did not have a
lowercased match to the \code{Preferred_Name} was used.
}
