% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_omop.R
\name{setup_omop}
\alias{setup_omop}
\title{Setup OMOP}
\usage{
setup_omop(
  conn,
  conn_fun = "pg13::local_connect()",
  target_schema = "omop_athena_nci",
  steps = c("drop_tables", "copy", "log", "indices", "constraints"),
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  checks = "",
  nci_version,
  log_schema = "public",
  log_table = "setup_nci_omop_log",
  output_folder,
  owl_folder = "/Users/mpatel/terminology/NCIT",
  neo4j_folder = "/Users/mpatel/Desktop/NCIt/neo4j",
  omop_folder = "/Users/mpatel/Desktop/NCIt/omop"
)
}
\description{
Instantiate a Postgres OMOP Vocabulary
database from the csvs created by the \code{process_owl_to_omop}
function.
}
